requirements.txt
scikit-learn
numpy
matplotlib
seaborn
xgboost
pandas
fpdf2
fastapi
uvicorn
